# Add support to disable html escaping in json encoder