--- a/json/json_test.go
+++ b/json/json_test.go
@@ -5,6 +5,8 @@ import (
 	"io"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	structform "github.com/elastic/go-structform"
 	"github.com/elastic/go-structform/sftest"
 )
@@ -50,4 +52,53 @@ func testEncParseConsistent(
 				return parse(buf.Bytes(), to)
 			}
 		})
+}
+
+func TestStringEncoding(t *testing.T) {
+	type testCase struct {
+		in         string
+		htmlEscape bool
+		expected   string
+	}
+
+	cases := map[string]testCase{
+		"no escaping required": testCase{
+			in:       "hello world",
+			expected: `"hello world"`,
+		},
+		"json required escaping": testCase{
+			in:         `"hello \nworld"`,
+			htmlEscape: false,
+			expected:   `"\"hello \\nworld\""`,
+		},
+		"html escape with json required escaping": testCase{
+			in:         `"hello \nworld"`,
+			htmlEscape: true,
+			expected:   `"\"hello \\nworld\""`,
+		},
+		"html with html escaping enabled": testCase{
+			in:         "<hello>world</hello>",
+			htmlEscape: true,
+			expected:   `"\u003chello\u003eworld\u003c/hello\u003e"`,
+		},
+		"html without html escaping": testCase{
+			in:         "<hello>world</hello>",
+			htmlEscape: false,
+			expected:   `"<hello>world</hello>"`,
+		},
+	}
+
+	for name, test := range cases {
+		in, htmlEscape, expected := test.in, test.htmlEscape, test.expected
+
+		t.Run(name, func(t *testing.T) {
+			buf := &bytes.Buffer{}
+			visitor := NewVisitor(buf)
+			visitor.SetEscapeHTML(htmlEscape)
+			visitor.OnString(in)
+
+			actual := buf.String()
+			assert.Equal(t, expected, actual)
+		})
+	}
 }